requirements.txt
flask
flask-SQLAlchemy
werkzeug
telebot
requests